chore: moves useDebounce to shared
